Move pandas tests to their own file
